added multi selector for genres
