+++ src/components/MovieForm/MovieForm.jsx
@@ -4,16 +4,42 @@ import { useHistory } from "react-router-dom";
 
 import Button from "@mui/material/Button";
 import Paper from "@mui/material/Paper";
+import { useTheme } from "@mui/material/styles";
+import Box from "@mui/material/Box";
+import OutlinedInput from "@mui/material/OutlinedInput";
 import TextField from "@mui/material/TextField";
 import Grid from "@mui/material/Grid";
 import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
+import Chip from "@mui/material/Chip";
+
+const ITEM_HEIGHT = 48;
+const ITEM_PADDING_TOP = 8;
+const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250,
+    },
+  },
+};
+
+function getStyles(name, personName, theme) {
+  return {
+    fontWeight:
+      personName.indexOf(name) === -1
+        ? theme.typography.fontWeightRegular
+        : theme.typography.fontWeightMedium,
+  };
+}
 
 export default function MovieForm() {
   const dispatch = useDispatch();
   const history = useHistory();
+
+  const theme = useTheme();
 
   // save the genres in the store on page load for the dropdown
   useEffect(() => {
@@ -27,7 +53,7 @@ export default function MovieForm() {
     title: "",
     poster: "",
     description: "",
-    genres: "",
+    genres: [],
   });
 
   const handleChange = (e) => {
@@ -60,7 +86,7 @@ export default function MovieForm() {
           onChange={handleChange}
           onSubmit={handleSubmit}
           sx={{
-            maxWidth: "max-content",
+            maxWidth: 1050,
             p: 3,
             mx: "auto",
             backgroundColor: "rgb(24,24,24)",
@@ -68,7 +94,7 @@ export default function MovieForm() {
           }}
           elevation={3}
         >
-          <Grid container rowSpacing={2} columnSpacing={2}>
+          <Grid container item rowSpacing={2} columnSpacing={2}>
             <Grid item xs={12}>
               <FormControl fullWidth>
                 <TextField
@@ -109,6 +135,37 @@ export default function MovieForm() {
               <FormControl fullWidth>
                 <InputLabel id="dropdown-label">Genre</InputLabel>
                 <Select
+                  labelId="dropdown-label"
+                  multiple
+                  name="genres"
+                  value={newMovie.genres}
+                  onChange={handleChange}
+                  input={
+                    <OutlinedInput id="select-multiple-chip" label="Chip" />
+                  }
+                  renderValue={(selected) => (
+                    <Box sx={{ display: "flex", flexWrap: "wrap", gap: 0.5 }}>
+                      {selected.map((value, i) => (
+                        <Chip key={i} label={value} />
+                      ))}
+                    </Box>
+                  )}
+                  // MenuProps={MenuProps}
+                >
+                  {genres.map((genre) => (
+                    <MenuItem
+                      key={genre.id}
+                      value={genre.name}
+                      style={getStyles(genre.name, newMovie.genres, theme)}
+                    >
+                      {genre.name}
+                    </MenuItem>
+                  ))}
+                </Select>
+              </FormControl>
+              {/* <FormControl fullWidth>
+                <InputLabel id="dropdown-label">Genre</InputLabel>
+                <Select
                   required
                   labelId="dropdown-label"
                   id="dropdown"
@@ -123,7 +180,7 @@ export default function MovieForm() {
                     </MenuItem>
                   ))}
                 </Select>
-              </FormControl>
+              </FormControl> */}
               {/* <select name="genre_id">
               <option>--Select a Genre--</option>
               {genres.map((genre, i) => (
